format event log fields for jsx, pin esio version
Let me begin by praising erlang dependency management, that of erlang.mk or
rebar in particular. Good luck repeating a build. You'll need it. A lot.
rebar3 is good, but the plugin must be built with emq, and erlang.mk there
doesn't bother with locks.

With that out of the way, this commit
 - pins esio version to preserve my sanity
 - cleans up or reformats to ensure all fields in events logged to es are
 stored in a usable format
   - id is a bitstring that can't be compared easily. base64 encoded it
   - peername and from are tuples that can't be converted to json by jsx.
   Converted to maps that will be represented as JSON Objects
   - got a little paranoid with reason and ensured it turns into a json
   string no matter what.